remove deprectated np.float alias (#497)
* remove deprectated np wrapper

* added release history and updated backtick numbers

+++ openmmtools/testsystems.py
@@ -493,7 +493,7 @@ class TestSystem(object):
         self._system = openmm.System()
 
         # Store positions.
-        self._positions = unit.Quantity(np.zeros([0, 3], np.float), unit.nanometers)
+        self._positions = unit.Quantity(np.zeros([0, 3], float), unit.nanometers)
 
         # Empty topology.
         self._topology = app.Topology()
